kubernetes_py_client/requirements.txt
adal==1.1.0
asn1crypto==0.24.0
cachetools==2.1.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
google-auth==1.5.1
idna==2.7
kubernetes==7.0.0
oauthlib==2.1.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyJWT==1.6.4
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
rsa==4.0
simplejson==3.16.0
six==1.11.0
urllib3==1.23
websocket-client==0.53.0
